실행은 되는데.. staticMesh는 이름이 겹치고, dynamicModel은 모델에 따라 되는게 있고 안되는게 있다..

+++ RocketEngine/RocketCommon/RocketTransform.cpp
@@ -328,7 +328,7 @@ namespace Rocket::Core
 
 	int RocketTransform::GetChildCount() const
 	{
-		return _children.size();
+		return (int)_children.size();
 	}
 
 	void RocketTransform::AddChild(RocketTransform* child)
